RT-13: Fixed GetPointingTo method

+++ src/camera/Camera.cs
@@ -45,12 +45,12 @@ namespace Raytracing
             pY.Mult((distance * Math.Tan(fOVY / 2)) / mag);
         }
 
-        public Vector3 GetPointingTo(int x, int y) {
+        public Vector3 GetPointingTo(double x, double y) {
             x -= resX / 2;
             y -= resY / 2;
 
-            int facX = x / (resX / 2);
-            int facY = y / (resY / 2);
+            double facX = x / (resX / 2);
+            double facY = y / (resY / 2);
 
             return Vector3.Add(Vector3.Add(pos, dir.Mult(distance)), Vector3.Add(pX.Mult(facX), pY.Mult(facY)));
         }
